@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Bank4Us.DataAccess.Core;
+using Microsoft.Data.SqlClient;
 
 namespace Bank4Us.DataAccess.Core
 {
