@@ -7,7 +7,6 @@ using Microsoft.AspNetCore.Mvc;
 using Bank4Us.BusinessLayer.Core;
 using Microsoft.Extensions.Logging;
 using Bank4Us.Common.Facade;
-using System.Web.Http;
 using System.Net.Http;
 using System.Text;
 
@@ -34,15 +33,15 @@ namespace  Bank4Us.ServiceApp.Controllers
         public IActionManager ActionManager { get { return _manager; } }
         public ILogger Logger { get { return _logger; } }
        
-        protected HttpResponseException LogException(Exception ex)
-        {
-            string errorMessage = LoggerHelper.GetExceptionDetails(ex);
-            _logger.LogError(LoggingEvents.SERVICE_ERROR, ex, errorMessage);
-            HttpResponseMessage message = new HttpResponseMessage();
-            message.Content = new StringContent(errorMessage);
-            message.StatusCode = System.Net.HttpStatusCode.ExpectationFailed;
-            throw new HttpResponseException(message);
-        }
+        //protected HttpResponseException LogException(Exception ex)
+        //{
+        //    string errorMessage = LoggerHelper.GetExceptionDetails(ex);
+        //    _logger.LogError(LoggingEvents.SERVICE_ERROR, ex, errorMessage);
+        //    HttpResponseMessage message = new HttpResponseMessage();
+        //    message.Content = new StringContent(errorMessage);
+        //    message.StatusCode = System.Net.HttpStatusCode.ExpectationFailed;
+        //    throw new HttpResponseException(message);
+        //}
 
     }
 }
