@@ -1,4 +1,4 @@
-﻿@page
+﻿@* @page
 @model ErrorModel
 @{
     ViewData["Title"] = "Error";
@@ -21,3 +21,4 @@
 <p>
     <strong>Development environment should not be enabled in deployed applications</strong>, as it can result in sensitive information from exceptions being displayed to end users. For local debugging, development environment can be enabled by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>, and restarting the application.
 </p>
+ *@
